% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.lad.R
\name{coef.lad}
\alias{coef.lad}
\title{Coefficients of a lad object}
\usage{
\method{coef}{lad}(object, ...)
}
\arguments{
\item{object}{An object of type lad}

\item{...}{further arguments passed to or from other methods}
}
\value{
Vector of coefficients of the lad object
}
\description{
Takes a lad object and returns its vector of coefficients
}
\details{
\code{coef.lad} returns the coefficients of the lad object.
}
