% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lad.R
\name{lad}
\alias{lad}
\title{Least absolute deviations regression}
\usage{
lad(x, y, ...)
}
\arguments{
\item{x}{A numeric vector of observations}

\item{y}{A numeric vector of observations}

\item{...}{further arguments passed to or from other methods}
}
\value{
An object of class "\code{lad}" is a list containing the following components:

\code{coefficients}    named regression coefficients

\code{fitted values}   vector of predicted values for the given x

\code{residuals}       vector of difference between actual and predicted values of y
}
\description{
Finds the values of beta0 and beta1 that minimize the sum of least absolute deviations for given
vectors x and y.
}
\details{
\code{lad} minimizes the sum of least absolute deviations and finds the corresponding
estimates of beta0 and beta1.
}
\examples{
v = lad(x = area$land, y = area$farm)
print(v)
plot(area$land, area$farm, main = "Farm Area vs Land Area", xlab = "Land Area (sq. miles)", ylab = "Farm Area (sq. miles)")
abline(reg = lm(area$farm ~ area$land), col = "red")
abline(reg = v, col = "blue")
legend("topleft", legend = c("lm()", "lad()"), col = c("red", "blue"), lty = c("solid", "solid"), cex = .9)
points(quantile(area$land), predict(v, quantile(area$land)), col = "green", pch = 16)


}
