% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.lad.R
\name{predict.lad}
\alias{predict.lad}
\title{Find the predicted values using the lad function}
\usage{
\method{predict}{lad}(object, new.x, ...)
}
\arguments{
\item{object}{An object of type lad}

\item{new.x}{A numeric vector}

\item{...}{further arguments passed to or from other methods}
}
\value{
Vector of lad's predictions at the x values in new.x
}
\description{
Takes a lad object and a numeric vector, and returns the predicted values from the lad function.
}
\details{
\code{predict.lad} returns the predicted values using the found values of beta0 and beta1.
}
