% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.lad.R
\name{print.lad}
\alias{print.lad}
\title{Print method for the lad class}
\usage{
\method{print}{lad}(x, ...)
}
\arguments{
\item{x}{A lad object}

\item{...}{further arguments passed to or from other methods}
}
\value{
Does not return anything; only prints to the console
}
\description{
Prints the vector of regression coefficients to the console.
}
\details{
\code{print.lad} prints the vector of coefficients to the console.
}
